Updated the Restaurant code to meet the requirements
- Treated the "UnknownError" case
- The members in the Cmd are checked in the same file by the IsValid() function
- Changed CreateRestaurantOP file
- Added some new ErrorCodes in Restaurant.cs
